// types/IFormData.ts
export interface IFormData {
  genero: string;
  nome: string;
  data_nascimento: string;
  telefone: string;
  cep: string;
  email: string;
  senha: string;
  imgUrl: string;
  cpf: string;
  tipoSanguineo?: string;
  doencasCronicas?: string;
  alergias?: string;
  crm?: string;
  especialidade?: string;
}